0.3.0.1       06/03/2018 Added some extra tests while hunting down why the weight value wasn't working.

+++ src/Filter/Filters.php
@@ -127,7 +127,7 @@ trait Filters {
 
 
   static function doRound($val, $settings, $helper = NULL) {
-    return round($val);
+    return (int) round($val);
   }
